% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{duration2time}
\alias{duration2time}
\title{Seconds to formatted time intervals}
\usage{
duration2time(seconds)
}
\arguments{
\item{seconds}{}
}
\value{
vecotr of string in 00:00:00 format
}
\description{
Seconds to formatted time intervals
}
