% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ms2kmh}
\alias{ms2kmh}
\title{Convert meters/s to km/h}
\usage{
ms2kmh(data)
}
\arguments{
\item{data}{fit file read with fit::read.fit()}
}
\value{
vector of speed in km/h
}
\description{
Convert meters/s to km/h
}
