% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plotly_cadence}
\alias{plotly_cadence}
\title{Plot plotly cadence}
\usage{
plotly_cadence(data, span = 0.007, black = FALSE)
}
\arguments{
\item{data}{}

\item{span}{span value to loess function}
}
\value{
plotly object
}
\description{
Plot plotly cadence
}
