% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plotly_HeartRate}
\alias{plotly_HeartRate}
\title{Plot plotly heart rate based on garmin fenix 6x pro}
\usage{
plotly_HeartRate(data, title = "", opacity = 0.8, black = FALSE)
}
\arguments{
\item{data}{fit file read with fit::read.fit()}
}
\value{
plotly object
}
\description{
Plot plotly heart rate based on garmin fenix 6x pro
}
