% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plotly_speed}
\alias{plotly_speed}
\title{Plot plotly heart rate based on garmin fenix 6x pro}
\usage{
plotly_speed(data, title = "")
}
\arguments{
\item{data}{fit file read with fit::read.fit()}
}
\value{
plotly object
}
\description{
Plot plotly heart rate based on garmin fenix 6x pro
}
