% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plotly_zones}
\alias{plotly_zones}
\title{Plot heart_rate zones}
\usage{
plotly_zones(data, black = FALSE)
}
\arguments{
\item{data}{fit file read with fit::read.fit()}
}
\value{
plotly object
}
\description{
Plot heart_rate zones
}
