% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{seconds_to_time}
\alias{seconds_to_time}
\title{Seconds to H M S}
\usage{
seconds_to_time(seconds)
}
\arguments{
\item{time_in_seconds}{}
}
\value{
dataframe with H M S values
}
\description{
Seconds to H M S
}
