% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{set_colors}
\alias{set_colors}
\title{Get colors for cadence zones}
\usage{
set_colors(data)
}
\arguments{
\item{data}{fit file read with fit::read.fit()}
}
\value{
vector
}
\description{
Get colors for cadence zones
}
