% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_map}
\alias{plot_map}
\title{Plot leaflet map from garin fit file}
\usage{
plot_map(
  data,
  color = "red",
  opacity = 0.7,
  weight = 3,
  provider = "OpenStreetMap.Mapnik",
  quantile = 0.8
)
}
\arguments{
\item{data}{fit file read with fit::read.fit()}

\item{color}{route color}

\item{opacity}{route opacity}

\item{weight}{route weigth}

\item{provider}{map provider}
}
\value{
leaflet map
}
\description{
Plot leaflet map from garin fit file
}
